feat: scratchNote (wip)
