+++ src/components/sandbox/ScratchNoteComponent.astro
@@ -1,0 +1,7 @@
+---
+import ScratchNote from './react/ScratchNote.js';
+---
+
+<div>
+  <ScratchNote />
+</div>
