--- src/components/sandbox/react/NoteCard.tsx
@@ -1,30 +1,0 @@
-import {
-  Card,
-  CardContent,
-  CardDescription,
-  CardFooter,
-  CardHeader,
-  CardTitle,
-} from '@/components/ui/card';
-import { cn } from '@/lib/utils';
-
-type CardProps = React.ComponentProps<typeof Card>;
-
-function NoteCard({ className, ...props }: CardProps) {
-  return (
-    <Card className={cn('w-[380px]', className)} {...props}>
-      <CardHeader>
-        <CardTitle>Note</CardTitle>
-        <CardDescription>Take a new note</CardDescription>
-      </CardHeader>
-      <CardContent>
-        <p>Card Content</p>
-      </CardContent>
-      <CardFooter>
-        <p>Card Footer</p>
-      </CardFooter>
-    </Card>
-  );
-}
-
-export default NoteCard;
