+++ src/components/sandbox/react/ScratchNote.test.tsx
@@ -1,10 +1,10 @@
 import { describe, it, expect } from 'vitest';
 import { render, screen } from '@testing-library/react';
-import NoteCard from './NoteCard';
+import ScratchNote from './ScratchNote';
 
 describe('NoteCard', () => {
   it('renders the NoteCard component', () => {
-    render(<NoteCard />);
+    render(<ScratchNote />);
     expect(screen.getByText('Note')).toBeDefined();
     screen.debug(); // prints out the jsx in the App component unto the command line
   });
